Correct depth bias.
The bias towards higher numbers of reads was due to the way reads aligning to the same genomic position were handled. This was corrected by sorting not only by position, but also by an index which counts the number of reads at the current position from each file.

Fixed tabs.